Change StorageMinerActor.GetProvingWindow to return UIntArray instead of multiple returns. (#3646)
This is the only method utilizing multiple returns. Changing this will let us change the
dispatch.FunctionSignature to return a single value instead of array, and then propagate
this to the MessageReceipt. See #3568.

+++ internal/app/go-filecoin/porcelain/client.go
@@ -79,13 +79,13 @@ func listAsksFromActorResult(ctx context.Context, plumbing claPlubming, actorRes
 		return err
 	}
 
-	var asksIds []uint64
+	var asksIds []types.Uint64
 	if err := encoding.Decode(ret[0], &asksIds); err != nil {
 		return err
 	}
 
 	for _, id := range asksIds {
-		ask, err := getAskByID(ctx, plumbing, addr, id)
+		ask, err := getAskByID(ctx, plumbing, addr, uint64(id))
 		if err != nil {
 			return err
 		}
